adding new front end feature to project

+++ backend/src/main/java/com/hotel/reservation/model/Room.java
@@ -50,7 +50,7 @@ public class Room {
     /**
      * Price per night
      */
-    private BigDecimal pricePerNight;
+    private int pricePerNight;
 
     /**
      * Maximum number of guests allowed
